init utils.py and first methods in there
